Правки красоты для travis'а Попытка 2

+++ include/Investigation.cpp
@@ -4,7 +4,7 @@
 Investigation::Investigation(ExperimentType t, std::vector<int64_t>& sizes) {
   Investigation::type = t;
   int id = 0;
-  for (long long i=log2(sizes[0]/2); pow(2, i) < 3*sizes[sizes.size()-1]/2; i++)
+  for (int64_t i=log2(sizes[0]/2); pow(2, i) < 3*sizes[sizes.size()-1]/2; i++)
   {
     id++;
     Experiment e(id, type, pow(2, i));
