+++ include/header.hpp
@@ -3,5 +3,5 @@
 #ifndef INCLUDE_HEADER_HPP_
 #define INCLUDE_HEADER_HPP_
 #include <string>
-std::string sizeConvertor(long long value);
+std::string sizeConvertor(int64_t value);
 #endif // INCLUDE_HEADER_HPP_
